github-api-facade : directories restructured| fetch_feature, fetch_feature_route, get_followers methods implemented inside GitHubApiFacade class
