+++ src/server/github-api-facade/.gitignore
@@ -1,2 +1,2 @@
 test.py
-facade/__pycache__
+**/__pycache__
